move types to Types file
